# apps/api/docs/swagger/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "snapdone Go API. 인증 credential은 JWT가 아닌 opaque 서버 세션 토큰.",
        "title": "snapdone API",
        "contact": {},
        "version": "0.1"
    },
    "basePath": "/",
    "paths": {
...
        "/v1/auth/capabilities": {
            "get": {
                "description": "서버에 설정 · 구현된 로그인 provider. 인증이 비활성이면 503이다.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "auth"
                ],
                "summary": "로그인 수단 목록",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/httpserver.CapabilitiesResponse"
                        }
                    },
                    "503": {
                        "description": "인증 비활성 (provider_unavailable)",
                        "schema": {
                            "$ref": "#/definitions/httpserver.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/v1/auth/exchange": {
            "post": {
                "description": "콜백이 넘긴 result code를 앱 verifier · state와 함께 한 번만 세션으로 바꾼다.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "oauth"
                ],
                "summary": "로그인 결과 교환",
                "parameters": [
                    {
                        "description": "result code · 앱 PKCE verifier · 앱 state",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/httpserver.ExchangeRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/httpserver.LoginResponse"
                        }
                    },
                    "400": {
                        "description": "잘못된 요청 · 틀린 proof · 만료 · 재사용 (invalid_callback)",
                        "schema": {
                            "$ref": "#/definitions/httpserver.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "내부 오류 (provider_unavailable)",
                        "schema": {
                            "$ref": "#/definitions/httpserver.ErrorResponse"
                        }
                    },
                    "503": {
                        "description": "인증 · OAuth 비활성 (provider_unavailable)",
                        "schema": {
                            "$ref": "#/definitions/httpserver.ErrorResponse"
                        }
                    }
                }
            }
        },
...
        "/v1/auth/logout": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "이 credential의 세션만 취소한다(root면 child 포함). 모르는 · 만료된 credential도 204다.",
                "tags": [
                    "auth"
                ],
                "summary": "로그아웃",
                "responses": {
                    "204": {
                        "description": "No Content"
                    },
                    "401": {
                        "description": "Authorization 헤더 형식 오류 (session_expired)",
                        "schema": {
                            "$ref": "#/definitions/httpserver.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "내부 오류 (provider_unavailable)",
                        "schema": {
                            "$ref": "#/definitions/httpserver.ErrorResponse"
                        }
                    },
                    "503": {
                        "description": "인증 비활성 (provider_unavailable)",
                        "schema": {
                            "$ref": "#/definitions/httpserver.ErrorResponse"
                        }
                    }
                }
            }
        },
...
        "/v1/auth/session": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Bearer credential의 세션을 확인하고 idle 만료를 연장한다. 만료 · 취소 · 모르는 credential은 401이다.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "auth"
                ],
                "summary": "현재 세션",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/httpserver.SessionResponse"
                        }
                    },
                    "401": {
                        "description": "credential 없음 · 만료 · 취소 (session_expired)",
                        "schema": {
                            "$ref": "#/definitions/httpserver.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "내부 오류 (provider_unavailable)",
                        "schema": {
                            "$ref": "#/definitions/httpserver.ErrorResponse"
                        }
                    },
                    "503": {
                        "description": "인증 비활성 (provider_unavailable)",
                        "schema": {
                            "$ref": "#/definitions/httpserver.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/v1/onboarding": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "내 온보딩 단계와 사용 목적. mobile과 web이 같은 진행에서 이어 간다.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "onboarding"
                ],
                "summary": "온보딩 진행",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/httpserver.OnboardingResponse"
                        }
                    },
                    "401": {
                        "description": "credential 없음 · 만료 · 취소 (session_expired)",
                        "schema": {
                            "$ref": "#/definitions/httpserver.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "내부 오류 (provider_unavailable)",
                        "schema": {
                            "$ref": "#/definitions/httpserver.ErrorResponse"
                        }
                    },
                    "503": {
                        "description": "인증 비활성 (provider_unavailable)",
                        "schema": {
                            "$ref": "#/definitions/httpserver.ErrorResponse"
                        }
                    }
                }
            },
            "put": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "단계(intro · purpose · first-image)와 사용 목적을 저장한다. 목적은 first-image에서만 있고 빈 목록은 건너뜀이다.\nunsure는 혼자만 고를 수 있다. 같은 단계를 다시 저장하거나 한 단계 앞으로만 갈 수 있고, 온보딩을 마친 뒤에는 바꿀 수 없다.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "onboarding"
                ],
                "summary": "온보딩 진행 저장",
                "parameters": [
                    {
                        "description": "저장할 진행",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/httpserver.OnboardingRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/httpserver.OnboardingResponse"
                        }
                    },
                    "400": {
                        "description": "계약 밖의 진행 (invalid_onboarding)",
                        "schema": {
                            "$ref": "#/definitions/httpserver.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "credential 없음 · 만료 · 취소 (session_expired)",
                        "schema": {
                            "$ref": "#/definitions/httpserver.ErrorResponse"
                        }
                    },
                    "409": {
                        "description": "이미 마친 온보딩 (onboarding_complete) · 건너뛰거나 되돌아가는 단계 (onboarding_out_of_order)",
                        "schema": {
                            "$ref": "#/definitions/httpserver.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "내부 오류 (provider_unavailable)",
                        "schema": {
                            "$ref": "#/definitions/httpserver.ErrorResponse"
                        }
                    },
                    "503": {
                        "description": "인증 비활성 (provider_unavailable)",
                        "schema": {
                            "$ref": "#/definitions/httpserver.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/v1/processing-jobs": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "사진 한 장(multipart 필드 image)으로 처리 작업을 만들고 바로 돌아온다. 결과는 작업 조회로 받는다.\nJPEG · PNG · GIF · WebP만 받고 7,500,000 byte까지다. 사진은 저장하지 않는다.",
                "consumes": [
                    "multipart/form-data"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "processing"
                ],
                "summary": "사진 처리 시작",
                "parameters": [
                    {
                        "type": "file",
                        "description": "사진 한 장",
                        "name": "image",
                        "in": "formData",
                        "required": true
                    }
                ],
                "responses": {
                    "202": {
                        "description": "Accepted",
                        "schema": {
                            "$ref": "#/definitions/httpserver.ProcessingJobResponse"
                        }
                    },
                    "400": {
                        "description": "multipart가 아니거나 image 필드 없음 (invalid_image)",
                        "schema": {
                            "$ref": "#/definitions/httpserver.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "credential 없음 · 만료 · 취소 (session_expired)",
                        "schema": {
                            "$ref": "#/definitions/httpserver.ErrorResponse"
                        }
                    },
                    "413": {
                        "description": "상한 초과 (image_too_large)",
                        "schema": {
                            "$ref": "#/definitions/httpserver.ErrorResponse"
                        }
                    },
                    "415": {
                        "description": "지원하지 않는 형식 (unsupported_image)",
                        "schema": {
                            "$ref": "#/definitions/httpserver.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "내부 오류 (provider_unavailable)",
                        "schema": {
                            "$ref": "#/definitions/httpserver.ErrorResponse"
                        }
                    },
                    "503": {
                        "description": "처리 비활성 (provider_unavailable)",
                        "schema": {
                            "$ref": "#/definitions/httpserver.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/v1/processing-jobs/{jobId}": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "내 처리 작업의 상태와, 완료됐다면 결과. 다른 사용자의 작업은 없는 작업과 같다.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "processing"
                ],
                "summary": "사진 처리 작업",
                "parameters": [
                    {
                        "type": "string",
                        "description": "처리 시작이 돌려준 jobId",
                        "name": "jobId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/httpserver.ProcessingJobResponse"
                        }
                    },
                    "401": {
                        "description": "credential 없음 · 만료 · 취소 (session_expired)",
                        "schema": {
                            "$ref": "#/definitions/httpserver.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "없는 작업 (job_not_found)",
                        "schema": {
                            "$ref": "#/definitions/httpserver.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "내부 오류 (provider_unavailable)",
                        "schema": {
                            "$ref": "#/definitions/httpserver.ErrorResponse"
                        }
                    },
                    "503": {
                        "description": "처리 비활성 (provider_unavailable)",
                        "schema": {
                            "$ref": "#/definitions/httpserver.ErrorResponse"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
...
        "httpserver.CapabilitiesResponse": {
            "type": "object",
            "properties": {
                "providers": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    },
                    "example": [
                        "google"
                    ]
                }
            }
        },
        "httpserver.ErrorResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string",
                    "enum": [
                        "session_expired",
                        "provider_unavailable",
                        "invalid_callback",
                        "invalid_image",
                        "image_too_large",
                        "unsupported_image",
                        "job_not_found",
                        "invalid_onboarding",
                        "onboarding_complete",
                        "onboarding_out_of_order"
                    ],
                    "example": "session_expired"
                }
            }
        },
        "httpserver.ExchangeRequest": {
            "type": "object",
            "required": [
                "code",
                "state",
                "verifier"
            ],
            "properties": {
                "code": {
                    "type": "string",
                    "example": "one-time-result-code"
                },
                "state": {
                    "type": "string",
                    "example": "base64url-app-state-43-to-128-chars"
                },
                "verifier": {
                    "type": "string",
                    "example": "base64url-app-verifier-43-to-128-chars"
                }
            }
        },
...
        "httpserver.LoginResponse": {
            "type": "object",
            "properties": {
                "credential": {
                    "type": "string",
                    "example": "opaque-session-credential"
                },
                "session": {
                    "$ref": "#/definitions/httpserver.SessionResponse"
                }
            }
        },
...
        "httpserver.OnboardingRequest": {
            "type": "object",
            "required": [
                "step"
            ],
            "properties": {
                "purposes": {
                    "type": "array",
                    "items": {
                        "type": "string",
                        "enum": [
                            "food",
                            "shopping",
                            "travel",
                            "events",
                            "receipt",
                            "foreign-language",
                            "work",
                            "unsure"
                        ]
                    },
                    "example": [
                        "food",
                        "receipt"
                    ]
                },
                "step": {
                    "type": "string",
                    "enum": [
                        "intro",
                        "purpose",
                        "first-image"
                    ],
                    "example": "first-image"
                }
            }
        },
        "httpserver.OnboardingResponse": {
            "type": "object",
            "properties": {
                "purposes": {
                    "type": "array",
                    "items": {
                        "type": "string",
                        "enum": [
                            "food",
                            "shopping",
                            "travel",
                            "events",
                            "receipt",
                            "foreign-language",
                            "work",
                            "unsure"
                        ]
                    },
                    "example": [
                        "food",
                        "receipt"
                    ]
                },
                "step": {
                    "type": "string",
                    "enum": [
                        "intro",
                        "purpose",
                        "first-image",
                        "complete"
                    ],
                    "example": "first-image"
                }
            }
        },
        "httpserver.ProcessingFactResponse": {
            "type": "object",
            "properties": {
                "label": {
                    "type": "string",
                    "example": "날짜"
                },
                "value": {
                    "type": "string",
                    "example": "8월 20일 19시"
                }
            }
        },
        "httpserver.ProcessingJobResponse": {
            "type": "object",
            "properties": {
                "jobId": {
                    "type": "string",
                    "example": "4f1c2a9e-0000-4000-8000-000000000000"
                },
                "result": {
                    "$ref": "#/definitions/httpserver.ProcessingResultResponse"
                },
                "status": {
                    "type": "string",
                    "enum": [
                        "running",
                        "completed",
                        "failed"
                    ],
                    "example": "running"
                }
            }
        },
        "httpserver.ProcessingResultResponse": {
            "type": "object",
            "properties": {
                "category": {
                    "type": "string",
                    "enum": [
                        "place",
                        "event",
                        "receipt",
                        "foreign_text",
                        "shopping",
                        "work",
                        "other"
                    ],
                    "example": "event"
                },
                "confidence": {
                    "type": "string",
                    "enum": [
                        "high",
                        "medium",
                        "low"
                    ],
                    "example": "high"
                },
                "facts": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/httpserver.ProcessingFactResponse"
                    }
                },
                "suggestedAction": {
                    "type": "string",
                    "enum": [
                        "save_place",
                        "add_to_calendar",
                        "record_expense",
                        "translate",
                        "none"
                    ],
                    "example": "add_to_calendar"
                }
            }
        },
        "httpserver.SessionResponse": {
            "type": "object",
            "properties": {
                "expiresAt": {
                    "type": "string",
                    "format": "date-time",
                    "example": "2026-10-01T00:00:00Z"
                },
                "onboardingStep": {
                    "type": "string",
                    "enum": [
                        "intro",
                        "purpose",
                        "first-image",
                        "complete"
                    ],
                    "example": "intro"
                },
                "user": {
                    "$ref": "#/definitions/httpserver.UserResponse"
                }
            }
        },
        "httpserver.UserResponse": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "string",
                    "example": "4f1c2a9e-0000-4000-8000-000000000000"
                }
            }
        }
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "\"Bearer \u003copaque session credential\u003e\" 형식. 서버만 해석할 수 있는 세션 토큰.",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}